requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
aiohttp-jinja2==1.2.0
alabaster==0.7.12
aniso8601==7.0.0
ansible==2.8.4
apispec==2.0.2
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
asyncmock==0.4.1
asyncpg==0.20.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
bcrypt==3.1.7
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.4
cryptography==2.7
docutils==0.15.2
firebase-admin==2.18.0
filtrark==0.9.0
flasgger==0.9.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
gevent==1.4.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.3.3
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.23.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
imagesize==1.1.0
importlib-metadata==0.20
injectark==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
lazy-object-proxy==1.4.2
livereload==2.6.1
MarkupSafe==1.1.1
marshmallow==3.0.3
mccabe==0.6.1
migrark==0.4.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.1
mypy==0.720
mypy-extensions==0.4.1
openapi-spec-validator==0.2.8
packaging==19.1
paramiko==2.6.0
pathlib==1.0.1
pathtools==0.1.2
pip-review==1.0
pluggy==0.12.0
port-for==0.4
protobuf==3.9.1
psycopg2==2.8.4
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
pytest-aiohttp==0.3.0
pytest-asyncio==0.10.0
pytest-cov==2.7.1
python-rapidjson==0.9.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
six==1.12.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinx-autobuild==0.7.1
sphinx-bootstrap-theme==0.7.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
tenark==0.6.0
tornado==6.0.3
typed-ast==1.4.0
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
watchdog==0.9.0
wcwidth==0.1.7
Werkzeug==0.15.6
wrapt==1.11.2
zipp==0.6.0
